@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   PUBLIC_URL: 'http://localhost:3000',
   API_URL: 'http://localhost:3000',
   APP_PORT: '3000',
