@@ -1,4 +1,9 @@
-module.exports = {
+import path from 'path';
+import { fileURLToPath } from 'url';
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
+
+export default {
   transform: {},
   rootDir: '../',
   roots: ['<rootDir>/src/', '<rootDir>/spec/'],
