@@ -1,4 +1,4 @@
-const connectServer = require('./connectServer');
-const { APP_PORT } = require('./config');
+import connectServer from './connectServer.js';
+import { APP_PORT } from './config/index.js';
 
-connectServer.start(APP_PORT);
+await connectServer.start(APP_PORT).catch((e) => console.log(e));
