@@ -1,12 +1,12 @@
-const OAuth2Server = require('oauth2-server');
+import OAuth2Server from 'oauth2-server';
 
-const OauthModel = require('./oauth-model');
+import OauthModel from './oauth-model.js';
 
 const FIVE_MINUTES_IN_SECONDS = 5 * 60;
 const ONE_MONTH_IN_SECONDS = 60 * 60 * 24 * 30;
 const ONE_YEAR_IN_SECONDS = ONE_MONTH_IN_SECONDS * 12;
 
-module.exports = new OAuth2Server({
+export default new OAuth2Server({
   model: OauthModel,
   authorizationCodeLifetime: FIVE_MINUTES_IN_SECONDS,
   accessTokenLifetime: ONE_MONTH_IN_SECONDS,
