@@ -1,5 +1,5 @@
-import * as React from 'react';
-import { withRouter } from 'react-router-dom';
+import React from 'react';
+import { useHistory } from "react-router-dom";
 import { withStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
 import DeveloperModeIcon from '@material-ui/icons/DeveloperMode';
@@ -19,40 +19,35 @@ const styles = {
   },
 };
 
-class NoFound404Page extends React.PureComponent {
-  goToHome() {
-    const { history } = this.props;
+const NoFound404Page = function ({ classes }) {
+  const goToHome = function () {
+    const history = useHistory();
     setTimeout(history.push(`/home`), 1000);
-  }
+  };
 
-  render() {
-    const { classes } = this.props;
-
-    return (
-      <Card className={classes.card}>
-        <CardContent>
-          <Typography variant="h5" component="h2">
-            <DeveloperModeIcon />
-            {'  '} Page not Found
-          </Typography>
-          <Typography variant="body1" component="p">
-            It seems that you have landed here by mistake. No problem, you can
-            start from home again.
-          </Typography>
-        </CardContent>
-        <CardActions>
-          <Button size="small" onClick={() => this.goToHome()}>
-            Go to home
-          </Button>
-        </CardActions>
-      </Card>
-    );
-  }
-}
+  return (
+    <Card className={classes.card}>
+      <CardContent>
+        <Typography variant="h5" component="h2">
+          <DeveloperModeIcon />
+          {'  '} Page not Found
+        </Typography>
+        <Typography variant="body1" component="p">
+          It seems that you have landed here by mistake. No problem, you can
+          start from home again.
+        </Typography>
+      </CardContent>
+      <CardActions>
+        <Button size="small" onClick={() => goToHome()}>
+          Go to home
+        </Button>
+      </CardActions>
+    </Card>
+  );
+};
 
 NoFound404Page.propTypes = {
   classes: PropTypes.instanceOf(Object).isRequired,
-  history: PropTypes.instanceOf(Object).isRequired,
 };
 
-export default withRouter(withStyles(styles)(NoFound404Page));
+export default withStyles(styles)(NoFound404Page);
