@@ -6,8 +6,8 @@ import CardContent from '@material-ui/core/CardContent';
 import { withStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
 import PropTypes from 'prop-types'; // ES6
-import * as React from 'react';
-import { withRouter } from 'react-router-dom';
+import React, { useState, useEffect } from 'react';
+import { useHistory, useLocation } from "react-router-dom";
 import Button from '@material-ui/core/Button';
 import oauthService from '../../services/oauth-service';
 import UserService from '../../services/user-service';
@@ -28,23 +28,20 @@ const styles = {
   },
 };
 
-class OAuthAuthorizePage extends React.PureComponent {
-  constructor(props) {
-    super(props);
-    this.state = {
-      applicationError: null,
-      authorizationError: false,
-      application: null,
-      currentUser: null,
-      loading: true,
-    };
-    this.onUserLoggedIn = this.onUserLoggedIn.bind(this);
-    this.confirmAuthorization = this.confirmAuthorization.bind(this);
-  }
+const OAuthAuthorizePage = function ({ classes }) {
 
-  componentDidMount () {
-    const { location, history } = this.props;
-    const urlParams = new URLSearchParams(location.search);
+  const [applicationError, setApplicationError] = useState(null);
+  const [authorizationError, setAuthorizationError] = useState(false);
+  const [application, setApplication] = useState(null);
+  const [currentUser, setCurrentUser] = useState(null);
+  const [loading, setLoading] = useState(true);
+
+
+  const location = useLocation();
+  const history = useHistory();
+  const urlParams = new URLSearchParams(location.search);
+
+  useEffect(() => {
     const params = Object.fromEntries(urlParams);
 
     Promise.all([
@@ -54,169 +51,149 @@ class OAuthAuthorizePage extends React.PureComponent {
       }),
       UserService.getCurrentUserAsync(),
     ])
+      // eslint-disable-next-line no-shadow
       .then(([application, currentUser]) => {
         console.log({ application, currentUser });
 
         let error;
         switch (params.response_type) {
           case 'code':
-            break
+            break;
           case 'token':
             if (!application.allowImplicitGrant)
-              error = new Error('Implicit grant not allowed for this application')
-            break
+              error = new Error('Implicit grant not allowed for this application');
+            break;
           default:
-            error = new Error('Invalid response_type parameter')
+            error = new Error('Invalid response_type parameter');
         }
         if (error) {
-          this.setState({
-            loading: false,
-            applicationError: error,
-          });
+          setLoading(false);
+          setApplicationError(error);
 
-          return
+          return;
         }
 
-        this.setState({
-          applicationError: null,
-          application,
-          currentUser,
-          loading: false,
-        });
+        setApplicationError(null);
+        setApplication(application);
+        setCurrentUser(currentUser);
+        setLoading(false);
       })
       .catch((err) => {
         console.error(err);
         if (err.code === 141) {
-          this.setState({
-            loading: false,
-            applicationError: err,
-          });
+          setLoading(false);
+          setApplicationError(err);
         } else {
           history.push('/');
         }
       });
-  }
+  }, []);
 
-  onUserLoggedIn (currentUser) {
-    this.setState({ currentUser });
-  }
+  // eslint-disable-next-line no-shadow
+  const onUserLoggedIn = function (currentUser) {
+    setCurrentUser(currentUser);
+  };
 
-  async confirmAuthorization () {
-    const { location } = this.props;
-    const urlParams = new URLSearchParams(location.search);
+  const confirmAuthorization = async function () {
     const params = Object.fromEntries(urlParams);
     try {
       window.location.href = await oauthService.authorize(params);
     } catch (err) {
       console.log({ err });
-      this.setState({ authorizationError: true });
+      setAuthorizationError(true);
     };
-  }
+  };
 
-  render () {
-    const { classes } = this.props;
-    const {
-      loading,
-      application,
-      applicationError,
-      currentUser,
-      authorizationError,
-    } = this.state;
-    const { location } = this.props;
-    const urlParams = new URLSearchParams(location.search);
 
-    return (
-      <Card className={classes.card}>
-        <CardContent>
-          {loading && (
-            <div
-              style={{ width: '100%', display: 'flex', flexDirection: 'row' }}
+  return (
+    <Card className={classes.card}>
+      <CardContent>
+        {loading && (
+          <div
+            style={{ width: '100%', display: 'flex', flexDirection: 'row' }}
+          >
+            <CircularProgress className={classes.progress} />
+          </div>
+        )}
+        {!loading && applicationError && (
+          <Typography color="error">
+            {applicationError.message}. If you are the developer, please
+            ensure your OAuth flow is setup correctly.
+          </Typography>
+        )}
+        {!loading && application && (
+          <>
+            <Typography gutterBottom variant="h4" component="h1">
+              Connect
+            </Typography>
+            <Typography gutterBottom variant="h5" component="h2">
+              {application.name}
+            </Typography>
+            <Typography
+              gutterBottom
+              variant="body1"
+              color="textPrimary"
+              paragraph
             >
-              <CircularProgress className={classes.progress} />
-            </div>
-          )}
-          {!loading && applicationError && (
-            <Typography color="error">
-              {applicationError.message}. If you are the developer, please
-              ensure your OAuth flow is setup correctly.
+              {application.name} would like to send data to Connect in your
+              name. They will not have access to private data regarding your
+              Connect account. You will be able to revoke this access any time
+              from your Connect dashboard.
             </Typography>
-          )}
-          {!loading && application && (
-            <>
-              <Typography gutterBottom variant="h4" component="h1">
-                Connect
-              </Typography>
-              <Typography gutterBottom variant="h5" component="h2">
-                {application.name}
-              </Typography>
-              <Typography
-                gutterBottom
-                variant="body1"
-                color="textPrimary"
-                paragraph
-              >
-                {application.name} would like to send data to Connect in your
-                name. They will not have access to private data regarding your
-                Connect account. You will be able to revoke this access any time
-                from your Connect dashboard.
-              </Typography>
-              {currentUser ? (
-                <>
-                  <Typography
-                    variant="body1"
-                    color="textSecondary"
-                    component="p"
+            {currentUser ? (
+              <>
+                <Typography
+                  variant="body1"
+                  color="textSecondary"
+                  component="p"
+                >
+                  By clicking on &quot;Connect&quot;, you authorize{' '}
+                  {application.name} to send data in your name to the Connect
+                  system.
+                </Typography>
+                <CardActions>
+                  <Button
+                    size="large"
+                    color="primary"
+                    onClick={() => confirmAuthorization()}
                   >
-                    By clicking on &quot;Connect&quot;, you authorize{' '}
-                    {application.name} to send data in your name to the Connect
-                    system.
-                  </Typography>
-                  <CardActions>
-                    <Button
-                      size="large"
-                      color="primary"
-                      onClick={this.confirmAuthorization}
-                    >
-                      Connect to {application.name}
-                    </Button>
-                  </CardActions>
-                </>
-              ) : (
-                <>
-                  <Typography
-                    variant="body1"
-                    color="textSecondary"
-                    component="p"
-                  >
-                    You must log in using your Connect account in order to link
-                    it to {application.name}. If you do not have a Connect
-                    account yet, you can create it directly from here.
-                  </Typography>
-                  <LoginActions
-                    onUserLoggedIn={this.onUserLoggedIn}
-                    redirectPath={`/authorize?${urlParams.toString()}`}
-                  />
-                </>
-              )}
-            </>
-          )}
-          {!loading && authorizationError && (
-            <Typography color="error">
-              Authorization could not be completed, please try again later. If
-              you&apos;re the app&apos;s developer, you might want to double
-              check your integration.
-            </Typography>
-          )}
-        </CardContent>
-      </Card>
-    );
-  }
-}
-
-OAuthAuthorizePage.propTypes = {
-  location: PropTypes.object.isRequired,
-  classes: PropTypes.instanceOf(Object).isRequired,
-  history: PropTypes.instanceOf(Object).isRequired,
+                    Connect to {application.name}
+                  </Button>
+                </CardActions>
+              </>
+            ) : (
+              <>
+                <Typography
+                  variant="body1"
+                  color="textSecondary"
+                  component="p"
+                >
+                  You must log in using your Connect account in order to link
+                  it to {application.name}. If you do not have a Connect
+                  account yet, you can create it directly from here.
+                </Typography>
+                <LoginActions
+                  onUserLoggedIn={(user) => onUserLoggedIn(user)}
+                  redirectPath={`/authorize?${urlParams.toString()}`}
+                />
+              </>
+            )}
+          </>
+        )}
+        {!loading && authorizationError && (
+          <Typography color="error">
+            Authorization could not be completed, please try again later. If
+            you&apos;re the app&apos;s developer, you might want to double
+            check your integration.
+          </Typography>
+        )}
+      </CardContent>
+    </Card>
+  );
 };
 
-export default withRouter(withStyles(styles)(OAuthAuthorizePage));
+OAuthAuthorizePage.propTypes = {
+  classes: PropTypes.instanceOf(Object).isRequired,
+};
+
+export default withStyles(styles)(OAuthAuthorizePage);
